feat(stream): widen extent byte offsets u32→u64 + max_extent_size 16 GiB flag
Extents were capped at 4 GiB by `ReadBytesReq.offset: u32`; `max_extent_size`
was hardcoded 3 GiB. Widen every extent byte position on the read+append path
to u64 so a single extent can exceed 4 GiB, and make the seal threshold a
16 GiB-default flag — fewer, larger extents = less manager/etcd metadata
pressure at scale. No deployment exists yet, so persisted formats change in
place (no migration; same-commit deploy enforced by WIRE v5 MIN=MAX).

Wire (extent_rpc binary hot path): AppendReq.commit, AppendResp.offset/end,
ReadBytesReq.offset/length, ReadBytesResp.end, CommitLengthResp.length →
u64 (+ manager's hand-mirrored ExtCommitLengthResp.length, byte-for-byte).
Persisted: ValuePointer 16→24 B, partition_rpc SstLocation.offset/len +
TableLocations.vp_offset, SST MetaBlock.vp_offset, TableMeta.offset/len,
StreamAllocExtentReq.seal_commit Option<u64>, MultiModifySplitReq.*_sealed_length,
GetRedirectResp.value_offset/value_len. WAL record val_len stays u32 (single
value ≤ 4 GB invariant); counts (in_flight, shard counts, intra-block offsets)
stay u32.

EC read chunking (coco P1): read_with_layout routes EC extents straight to
ec_subrange_read, bypassing the replicated path's 256 MiB chunking. At 16 GiB
/ K=3 a shard is ~5.33 GiB, so a full-shard read would overflow the response
frame's payload_len: u32. read_shard_from_addr now chunks internally at
read_chunk_bytes — the single EC choke point — so every MSG_READ_BYTES stays
under 4 GiB. (Unreachable at the old 3 GiB default; the 16 GiB default exposes
it.)

max_extent_size: set_max_extent_size_bytes (OnceLock, default 16 GiB, clamp
[1,64] GiB) read at both StreamClient::new_with_owner_epoch sites; new
`autumn-ps --max-extent-size-bytes` flag.

WIRE fingerprint v4→v5 (MIN=MAX=5, pre-R3 same-commit deploy).

Validated: cargo build --workspace green; rpc 30 / manager lib 163 / stream +
PS unit suites green (the 3 manager-integration recovery tests are pre-existing
contention-flaky — identical on HEAD baseline). Live isolated e2e (memory-mode
4-EN loopback cluster): put-stream 4.5 GiB into one log_stream extent (on-disk
4,831,907,398 B > u32::MAX) → get-stream sha256 MATCH (later chunk VPs at
offset > 4 GiB, served via u64 ReadBytesReq.offset) → kill -9 PS → recovery
replays SST (paged u64 vp_offset) + WAL (u64 record offset) → re-read MATCH.

+++ crates/manager/src/lib.rs
@@ -2716,7 +2716,7 @@ the manager binaries first (design §6: bump comes AFTER all members run the new
         state: &autumn_common::MetadataState,
         src_stream_id: u64,
         dst_stream_id: u64,
-        sealed_length: u32,
+        sealed_length: u64,
     ) -> Result<(MgrStreamInfo, Vec<MgrExtentInfo>), AppError> {
         let src = state
             .streams
@@ -2749,7 +2749,7 @@ the manager binaries first (design §6: bump comes AFTER all members run the new
             // stream alloc a fresh tail on init instead of sharing this one.
             if idx == src.extent_ids.len() - 1 && !ex.sealed {
                 ex.sealed = true;
-                ex.sealed_length = sealed_length as u64;
+                ex.sealed_length = sealed_length;
                 ex.avali = Self::all_bits(ex.replicates.len() + ex.parity.len());
                 // BUG2 trace (opt-in): split CoW-tail seal. A `sealed_length=0`
                 // here freezes a shared tail that may hold VP/SST-acked data →
@@ -2789,8 +2789,8 @@ the manager binaries first (design §6: bump comes AFTER all members run the new
         state: &autumn_common::MetadataState,
         survivor_stream_id: u64,
         victim_stream_id: u64,
-        survivor_sealed: u32,
-        victim_sealed: u32,
+        survivor_sealed: u64,
+        victim_sealed: u64,
         new_tail: MgrExtentInfo,
     ) -> Result<(MgrStreamInfo, Vec<MgrExtentInfo>), AppError> {
         let survivor = state
@@ -2819,7 +2819,7 @@ the manager binaries first (design §6: bump comes AFTER all members run the new
             // way so the seal persists.
             if !ex.sealed {
                 ex.sealed = true;
-                ex.sealed_length = survivor_sealed as u64;
+                ex.sealed_length = survivor_sealed;
                 ex.eversion += 1;
                 ex.avali = Self::all_bits(ex.replicates.len() + ex.parity.len());
                 modified_extents.push(ex);
@@ -2840,7 +2840,7 @@ the manager binaries first (design §6: bump comes AFTER all members run the new
             // (coco P1, CoW isolation).
             if idx == victim.extent_ids.len() - 1 && !ex.sealed {
                 ex.sealed = true;
-                ex.sealed_length = victim_sealed as u64;
+                ex.sealed_length = victim_sealed;
                 ex.avali = Self::all_bits(ex.replicates.len() + ex.parity.len());
             }
             modified_extents.push(ex);
@@ -2872,8 +2872,8 @@ the manager binaries first (design §6: bump comes AFTER all members run the new
         state: &autumn_common::MetadataState,
         survivor_stream_id: u64,
         victim_stream_id: u64,
-        survivor_sealed: u32,
-        victim_sealed: u32,
+        survivor_sealed: u64,
+        victim_sealed: u64,
     ) -> Result<(MgrStreamInfo, Vec<MgrExtentInfo>), AppError> {
         let survivor = state
             .streams
@@ -2899,7 +2899,7 @@ the manager binaries first (design §6: bump comes AFTER all members run the new
             // way so the seal persists.
             if !ex.sealed {
                 ex.sealed = true;
-                ex.sealed_length = survivor_sealed as u64;
+                ex.sealed_length = survivor_sealed;
                 ex.eversion += 1;
                 ex.avali = Self::all_bits(ex.replicates.len() + ex.parity.len());
                 modified_extents.push(ex);
@@ -2918,7 +2918,7 @@ the manager binaries first (design §6: bump comes AFTER all members run the new
             // (coco P1, CoW isolation).
             if idx == victim.extent_ids.len() - 1 && !ex.sealed {
                 ex.sealed = true;
-                ex.sealed_length = victim_sealed as u64;
+                ex.sealed_length = victim_sealed;
                 ex.avali = Self::all_bits(ex.replicates.len() + ex.parity.len());
             }
             modified_extents.push(ex);
@@ -3226,7 +3226,7 @@ the manager binaries first (design §6: bump comes AFTER all members run the new
         addr: &str,
         extent_id: u64,
         owner_epoch: i64,
-    ) -> Result<u32, AppError> {
+    ) -> Result<u64, AppError> {
         debug_assert!(
             owner_epoch > 0,
             "commit_length_on_node requires owner_epoch > 0; use probe_extent_on_node for fence-free probes"
@@ -3274,7 +3274,7 @@ the manager binaries first (design §6: bump comes AFTER all members run the new
         &self,
         addr: &str,
         extent_id: u64,
-    ) -> Result<u32, AppError> {
+    ) -> Result<u64, AppError> {
         let base = Self::normalize_endpoint(addr);
         let shard_ports = self.shard_ports_for_addr(&base);
         let routed = Self::shard_addr_for_extent(&base, &shard_ports, extent_id);
